watchlist and header changes
